backend/requirements.txt
gigachat
python-dotenv
fastapi